refactor : x버튼 svg로 변경 및 컴포넌트 디자인 수정
